requirements/req.txt
amqp==5.2.0
asgiref==3.8.1
asttokens==2.4.1
async-timeout==4.0.3
beautifulsoup4==4.12.3
billiard==4.2.0
celery==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
decorator==5.1.1
Django==5.0.7
django-celery-beat==2.6.0
django-timezone-field==7.0
djangorestframework==3.15.2
exceptiongroup==1.2.2
executing==2.0.1
idna==3.7
ipython==8.26.0
jedi==0.19.1
kombu==5.3.7
matplotlib-inline==0.1.7
parso==0.8.4
pexpect==4.9.0
prompt_toolkit==3.0.47
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.0.7
requests==2.32.3
six==1.16.0
soupsieve==2.5
sqlparse==0.5.1
stack-data==0.6.3
structlog==24.4.0
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
